requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

